web/requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.2
django-appconf==1.0.2
django-compressor==2.2
django-registration==3.0
django-sass-processor==0.7.1
djangorestframework==3.8.2
gunicorn==19.9.0
libsass==0.15.1
psycopg2-binary==2.7.5
pytz==2018.5
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
